requirements.txt
Django==2.0.2
djangorestframework==3.7.7
django_debug_toolbar==1.9.1
django-extensions==2.0.0
psycopg2==2.7.4
celery==4.1.0
requests==2.18.4
django-rest-swagger==2.1.2
pyOpenSSL==17.5.0
rsa==3.4.2
django-flat-responsive==2.0
django-admin-view-permission==1.5
django-admin-list-filter-dropdown==1.0.1
django-import-export==1.0.0
openpyxl==2.4.9
django-admin-rangefilter==0.3.1
django-redis==4.9.0
raven==6.6.0
flower
